Correct capstone clone outside builddir ##build
While current implementation of the `run_command` runs inside the
subdirectory of the meson.build file it doesn't have to work that way.

As per meson's own documentation for `run_command`:
> The command is run from an unspecified directory

It just so happens that muon (C implementation of meson) runs those
commands from the current working directory, which results in capstone
being cloned in the project root, which breaks `include_directories`
call for `capstone_includes`.

Properly specifying `capstone_path` and using `-C` argument for the
clone fixes the issue.